First pass of converting Flow annotations to JSDoc
